graph_rag_course/stage1_vector_rag/requirements.txt
langchain>=0.0.334
langchain-community>=0.0.21
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
pinecone-client>=2.2.4
python-dotenv>=1.0.0
wikipedia>=1.4.0
langgraph>=0.0.19
